BackEnd #fix : 클러스터 검색

+++ src/main/java/com/ssafy/BravoSilverLife/entity/BDCode.java
@@ -15,7 +15,7 @@ import javax.persistence.Table;
 @Table(name = "bd_code")
 public class BDCode {
     @Id
-    int code;
+    long code;
     String sido;
     String gugun;
     String name;
